+++ HISTORY.md
@@ -1,4 +1,11 @@
 # History
+## 0.0.4 (2023-07-16)
+- **VimeoBaseItem**: method **set_temp_data** (for use in Restricted Python where methods with the setter-decorator do not work because attributes cannot be changed)
+- **VimeoBaseItem**: new attribute *\_temp_attributes* for storing volatile "attributes" on the object and the methods needed:
+	- get_attribute
+	- get_attributes
+	- set_attribute
+
 ## 0.0.3 (2023-07-13)
 - **VimeoShocase** und **VimeoFolder**: property **nb_videos**
 - **VimeoData** changed property names